feat(reth-bench): support human-readable gas format in generate-big-block
Add support for K/M/G suffixes (e.g., '1G', '30M') to the --target-gas
parameter in generate-big-block, matching the existing format in
gas-limit-ramp's --target-gas-limit.

- Move parse_gas_limit to helpers.rs for shared use
- Update generate-big-block to use the shared parser
- Refactor gas-limit-ramp to import from helpers
